"""El que passa es que la hora la considera com un float, es un error de wx. He de canviar la forma de escriure la hora, fent servir un altre objecte, com per exemple, time_string | wxDateTime | wxTimeSpan | mx.DateTime | mx.DateTimeDelta"""
In this commit, we solve some errors.

1) dc.DrawText needs int arguments, not float

2) Solve image.SetOptionInt deprecation, now using SetOption with an integer argument

3) Substitute StockCursor for his updated version Cursor.

4) Substitute wx.ImageFromBitmap for his newer version wx.bitmap.ConvertToImage

5) Solve reference wx.GridTableMessage, now using wx.grid.GridTableMessage.

6) Solve wx.grid.GRIDTABLE_REQUEST_VIEW_GET_VALUES reference

7) Solve wx.grid.GridCellAttr() reference.

8) Remove column argument in tablecolumn.py, it was not necessary at the end. I have to check other modules now, because I remember I changed some of them related with this column argument.

+++ plugins/extendededitor/extendededitor/item.py
@@ -40,7 +40,7 @@ class RoundedGradedHeader(object):
         inset = self.radius / 2
         for i in range(size[1]):
             dc.SetPen(wx.Pen(wx.Colour(startColour[0], startColour[1], startColour[2])))
-            dc.DrawLine(inset, i, size[0] - inset * 2, i)
+            dc.DrawLine(int(inset), i, int(size[0] - inset * 2), i)
             if startColour[0] + jump < 255:
                 startColour[0] += jump
             if startColour[1] + jump < 255:
@@ -137,7 +137,7 @@ class RoundedGradedHeader(object):
         offset = 3
         if self.image is not None:
             top = self.calcCenter(self.image.GetHeight())
-            dc.DrawBitmap(self.image, offset, top, True)
+            dc.DrawBitmap(self.image, offset, int(top), True)
             offset += self.image.GetWidth() + offset
         if self.text is not None:
             top = self.calcCenter(dc.GetTextExtent(self.text)[1])
